refactor: manually handle zustand merges (#149)
* refactor: manually handle zustand merges

* refactor: remove console log

* refactor: cleanup pr

+++ src/stores/storage.ts
@@ -6,18 +6,34 @@ import _ from "lodash";
 
 export function createStorage<S>(): PersistStorage<S> {
   const db = createDb("zustand");
+
+  // If read takes too long, write can override it
+  // so we lock writing until read finishes.
+  let locked = true;
+
   return {
     getItem: async (key) => {
-      const value = await pRetry(() => db.getItem(key), {
-        retries: 3,
-      });
-      return value ? JSON.parse(value) : null;
+      try {
+        const value = await pRetry(() => db.getItem(key), {
+          retries: 5,
+        });
+        locked = false;
+        return value ? JSON.parse(value) : null;
+      } catch (err) {
+        console.error(err);
+        locked = false;
+        return null;
+      }
     },
     setItem: async (key, value) => {
-      await db.setItem(key, JSON.stringify(value));
+      if (!locked) {
+        await db.setItem(key, JSON.stringify(value));
+      }
     },
     removeItem: async (key) => {
-      await db.removeItem(key);
+      if (!locked) {
+        await db.removeItem(key);
+      }
     },
   };
 }
